feat: Refactor GachaSettingView UI and extract components
- Extract GachaItemEditForm from inline modal form to a separate widget for
  better modularity and readability.
- Extract GachaItemsSection widget to encapsulate the gacha items list and
  related controls, replacing previous inline UI code.
- Extract GachaTitleBar widget for the app bar title area with user name and
  subtitle inputs.
- Replace multiple BlocProvider.value usages with explicit generic types for
  clarity.
- Simplify and clean up GachaSettingView build method by using new components.
- Remove deprecated _buildEditForm, _buildTitleBar, and related UI helper
  methods from GachaSettingView.
- Maintain existing functionality including
